% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Subset a column of dataframes}
\usage{
select(x, var)
}
\arguments{
\item{x}{A dataframe}

\item{var}{Either a vector of column indices, or variable names or logicals}
}
\value{
A subset of a dataframe
}
\description{
Subset a column of dataframes
}
