% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select2}
\alias{select2}
\title{Select variables}
\usage{
select2(df, var)
}
\arguments{
\item{df}{A data frame.}

\item{var}{A character or integer vector with variable names or indices to select.}
}
\value{
The subsetted data frame.
}
\description{
Select variables in a data frame.
}
\examples{
select2(iris, 1:3)
}
